# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_fisker_zjump-backend_internal_model.BillPrice:
    properties:
      cluster:
        description: 集群
        type: string
      createdAt:
        type: string
      description:
        description: 描述
        type: string
      id:
        type: integer
      price:
        description: 单价
        type: number
      resource_type:
        description: 资源类型
        type: string
      scale:
        description: 规格，如 1:2, 1:4
        type: string
      updatedAt:
        type: string
      vendor:
        description: 云厂商
        type: string
    required:
    - vendor
    type: object
...
  github_com_fisker_zjump-backend_internal_model.QueryLog:
    properties:
      affectedRows:
        type: integer
      clientIp:
        type: string
      createdAt:
        type: string
      databaseName:
        type: string
      dbType:
        type: string
      errorMessage:
        type: string
      executionTimeMs:
        type: integer
      id:
        type: integer
      instanceId:
        type: integer
      instanceName:
        type: string
      queryContent:
        type: string
      queryType:
        type: string
      resultCount:
        type: integer
      resultPreview:
        type: string
      status:
        type: string
      userAgent:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
  github_com_fisker_zjump-backend_internal_model.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@keyops.com
    name: API Support
  description: KeyOps 基础设施管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: KeyOps API
  version: "2.0"
paths:
...
  /api/bill/price:
    get:
      consumes:
      - application/json
      description: 获取单价管理列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取单价列表
      tags:
      - bill
    post:
      consumes:
      - application/json
      description: 创建单价配置
      parameters:
      - description: 单价信息
        in: body
        name: price
        required: true
        schema:
          $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.BillPrice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 创建单价
      tags:
      - bill
  /api/bill/price/:id:
    put:
      consumes:
      - application/json
      description: 更新单价配置
      parameters:
      - description: 单价ID
        in: path
        name: id
        required: true
        type: string
      - description: 单价信息
        in: body
        name: price
        required: true
        schema:
          $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.BillPrice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 更新单价
      tags:
      - bill
  /api/bill/records:
    get:
      consumes:
      - application/json
      description: 获取账单明细列表，支持分页和筛选
      parameters:
      - description: 云厂商 (tencent/huawei-langgemap/huawei-bjlg)
        in: query
        name: vendor
        required: true
        type: string
      - description: '账单月份 (格式: 2024-01)'
        in: query
        name: month
        required: true
        type: string
      - description: 资源类型代码
        in: query
        name: resource_code
        type: string
      - description: 服务类型代码
        in: query
        name: service_code
        type: string
      - default: 1
        description: 页码，从1开始
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: "0"
        description: 是否从云厂商API查询 (0=本地, 1=远程)
        in: query
        name: remote
        type: string
      - default: "0"
        description: 是否计算费用 (0=否, 1=是)
        in: query
        name: with_amount
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取账单明细列表
      tags:
      - bill
...
  /api/bill/statistics:
    get:
      consumes:
      - application/json
      description: 获取当月总费用，用于前端展示饼图
      parameters:
      - description: 账单月份
        in: query
        name: month
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取费用统计
      tags:
      - bill
  /api/bill/summary:
    get:
      consumes:
      - application/json
      description: 获取指定云厂商和月份的账单汇总信息
      parameters:
      - description: 云厂商
        in: query
        name: vendor
        required: true
        type: string
      - description: 账单月份
        in: query
        name: month
        required: true
        type: string
      - default: "0"
        description: 是否从云厂商API查询
        in: query
        name: remote
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取月度账单汇总
      tags:
      - bill
  /api/bill/trend:
    get:
      consumes:
      - application/json
      description: 获取月度费用列表，用于前端展示折线图
      parameters:
      - description: 云厂商
        in: query
        name: vendor
        type: string
      - description: 年份
        in: query
        name: year
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取费用趋势
      tags:
      - bill
  /api/bill/trend/month:
    get:
      consumes:
      - application/json
      description: 获取折线图上x轴的月份列表，默认查询最近6个月
      parameters:
      - description: 年份
        in: query
        name: year
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取费用趋势月份列表
      tags:
      - bill
  /api/bill/vm:
    get:
      consumes:
      - application/json
      description: 获取虚拟机分摊账单（本地数据库，包含硬盘）
      parameters:
      - description: 云厂商
        in: query
        name: vendor
        required: true
        type: string
      - description: 账单月份
        in: query
        name: month
        required: true
        type: string
      - description: 分摊类型 (department/business)
        in: query
        name: split_type
        type: string
      - default: "0"
        description: 是否包含详情
        in: query
        name: with_detail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
      summary: 获取虚拟机分摊账单
      tags:
      - bill
...
  /api/dms/logs/queries:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 实例ID
        in: query
        name: instance_id
        type: integer
      - description: 数据库类型
        in: query
        name: db_type
        type: string
      - description: 查询类型
        in: query
        name: query_type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      - description: 搜索关键词
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.QueryLog'
                      type: array
                    total:
                      format: int64
                      type: integer
                  type: object
              type: object
      summary: 获取查询日志列表
      tags:
      - DMS
  /api/dms/logs/queries/:id:
    get:
      consumes:
      - application/json
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_fisker_zjump-backend_internal_model.QueryLog'
              type: object
      summary: 获取查询日志详情
      tags:
      - DMS
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"